fixed display for ugd/uugd

+++ fx/src/main/java/org/unigrid/janus/model/StakedBalanceModel.java
@@ -6,5 +6,5 @@ import lombok.Data;
 @Data
 @ApplicationScoped
 public class StakedBalanceModel {
-    private long stakedBalance;
+    private double stakedBalance;
 }
